+++ src/main/java/com/example/hospitalreservation/common/ErrorResponse.java
@@ -1,4 +1,4 @@
-package com.example.hospitalreservation.model;
+package com.example.hospitalreservation.common;
 
 public class ErrorResponse {
     private String message;
