+++ src/main/java/com/example/hospitalreservation/model/DeleteReservationResponse.java
@@ -1,5 +1,7 @@
 package com.example.hospitalreservation.model;
 
+import com.example.hospitalreservation.common.SuccessMessage;
+
 public class DeleteReservationResponse {
     private String message;
